RAD-381 Create RadiologyOrderServiceTest
